% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thslsfunction.R
\name{summary.tsls}
\alias{summary.tsls}
\title{Summary}
\usage{
\method{summary}{tsls}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\description{
Summary
}

