% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thslsfunction.R
\name{tsls.formula}
\alias{tsls.formula}
\title{formula}
\usage{
\method{tsls}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{PIB~INF+TIR|Cap+m2r}

\item{data}{the dataframe}

\item{...}{not used}
}
\description{
formula
}

