% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thslsfunction.R
\name{tsls}
\alias{tsls}
\title{method}
\usage{
tsls(x, ...)
}
\arguments{
\item{x}{a numeric design matrix for the model.}

\item{...}{not used}
}
\description{
method
}
\author{
Zaghdoudi Taha
}

